+++ libs/src/TAGbase/TAGrecoManager.hxx
@@ -81,15 +81,15 @@ public:
    //! TOE TGT tag
    const char*     GetTgtTag(int idx)            { return fTgtTag[idx].c_str();}
    //! TOE VTX tag cut
-   vector<int>     GetVtxTagCuts(string idx)     { return fVtxTagCuts[idx];  }
+   vector<size_t>  GetVtxTagCuts(string idx)     { return fVtxTagCuts[idx];  }
    //! TOE ITR tag cut
-   vector<int>     GetItrTagCuts(string idx)     { return fItrTagCuts[idx];  }
+   vector<size_t>  GetItrTagCuts(string idx)     { return fItrTagCuts[idx];  }
    //! TOE MSD tag cut
-   vector<int>     GetMsdTagCuts(string idx)     { return fMsdTagCuts[idx];  }
+   vector<size_t>  GetMsdTagCuts(string idx)     { return fMsdTagCuts[idx];  }
    //! TOE MSD 2D tag cut
-   vector<int>     GetMsd2TagCuts(string idx)    { return fMsd2TagCuts[idx]; }
+   vector<size_t>  GetMsd2TagCuts(string idx)    { return fMsd2TagCuts[idx]; }
    //! TOE TW tag cut
-   vector<int>     GetTwTagCuts(string idx)      { return fTwTagCuts[idx];   }
+   vector<size_t>  GetTwTagCuts(string idx)      { return fTwTagCuts[idx];   }
 
     //! Reconstruction parameters Getter
     //! Local reconstruction flag
@@ -245,11 +245,11 @@ private:
    
    // TOE parameters
    vector<string>                    fTgtTag;      ///< TOE TGT tag
-   map<string, vector<int>>          fVtxTagCuts;  ///< TOE VTX cuts
-   map<string, vector<int>>          fItrTagCuts;  ///< TOE ITR cuts
-   map<string, vector<int>>          fMsdTagCuts;  ///< TOE MSD cuts
-   map<string, vector<int>>          fMsd2TagCuts; ///< TOE MSD 2D cuts
-   map<string, vector<int>>          fTwTagCuts;   ///< TOE TW cuts
+   map<string, vector<size_t>>       fVtxTagCuts;  ///< TOE VTX cuts
+   map<string, vector<size_t>>       fItrTagCuts;  ///< TOE ITR cuts
+   map<string, vector<size_t>>       fMsdTagCuts;  ///< TOE MSD cuts
+   map<string, vector<size_t>>       fMsd2TagCuts; ///< TOE MSD 2D cuts
+   map<string, vector<size_t>>       fTwTagCuts;   ///< TOE TW cuts
 
    
    //! reconstruction parameter
